ovn-controller: Detect OVS datapath capabilities.
Automatically create an OVS Datapath record if none exists for the
current br-int datapath type.

Add a 'features' module to track which OVS features are available in
the datapath currently being used.  For now, only ct_zero_snat is
tracked, all other features are assumed to be on-par between all
datapaths.

A future commit will make use of the 'features' module to conditionally
program openflows based on available datapath features.

Acked-by: Mark D. Gray <mark.d.gray@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>