northd: I-P add explicit dependency on Load_Balancer_Group.
The commit that added I-P for northd missed to add this explicit
dependency.  This didn't really affect functionality because:
a. ovn-northd monitors the whole NB database
b. changes to the Load_Balancer_Group table mark rows in Logical_Switch
   table that refer to the updated LB groups as "updated".

However, it's probably better to make this dependency explicit having in
mind that real incremental processing might be implemented in the
future.

Fixes: 4597317f16d1 ("northd: Introduce incremental processing for northd")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>