ovn-northd: Don't trigger recompute for transactions that are in progress.
Most of the times ovsdb_idl_loop_commit_and_wait() will return -1
indicating that the newly initiated transaction is still in progress.

That's not an error condition and it shouldn't trigger a forced
recompute.  Moreover, arguably, we shouldn't clear tracked rows only
when the transaction is successful.  We should probably do it
unconditionally in every iteration, to avoid processing the same update
twice.

Fixes: 4597317f16d1 ("northd: Introduce incremental processing for northd")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>