acl-log: Log the direction (logical pipeline) of the matching ACL.
It's useful to differentiate between ingress and egress pipelines in the
ACL logs.  To achieve this we expand the "log()" logical action and pass
an optional direction.

This behavior change is implemented in a backwards compatible way such
that it doesn't break existing deployments even in the case when
ovn-northd gets updated before ovn-controller.

To achieve this, ovn-northd determines first if all chassis in the
cluster have been upgraded to a version that supports the new "log()"
action format.

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=1992641
Signed-off-by: Dumitru Ceara <dceara@redhat.com>