inc-proc-northd: Only run the engine when both NB and SB txns are available.
This is to avoid unnecessary forced recomputes.  E.g., when northd
updates the SB, adding a logical flow, the update message from SB is
received before the transaction reply, essentially forcing a full
recompute for every SB change.

Due to the fact that tracked changes can be accumulated and that
transactions can not be indefinitely in-progress we now clear tracked
changes when the engine actually can run.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>