northd, controller: Add timestamp column to MAC_Binding table
The new timestamp column in MAC_Binding is
populated with current time whenever the row is
created or the MAC address is updated.
This can be utilized by MAC binding aging mechanism,
when we can check if enough time has passed since the
creation/update.

Reported-at: https://bugzilla.redhat.com/2084668
Acked-By: Ihar Hrachyshka <ihrachys@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>