binding.c: Missing local_datapath update in runtime_data port_binding handler.
When handling port_binding changes, it is possible that new
local_datapaths are added, and the fields of the local_datapath, such as
localnet_port, external_ports, etc. need to be updated for the newly
added local_datapaths.

This problem doesn't happen in most cases because the changes that
trigger the new local_datapath addition usually trigger recomputes,
but it may not always be the case. If recompute is not triggered,
local_datapaths are not properly updated and will result in missing OVS
flows. This is more likely to happen when we delay patch interface
deletion in the future, or if we implement patch interface I-P.

Signed-off-by: Han Zhou <hzhou@ovn.org>