tests: Simplify ovn-northd specific tests.
Avoid having to add OVN_FOR_EACH_NORTHD for each test in ovn-northd.at.
We know that all tests in there should be run for all types of northd
and w/ or w/o parallelization enabled.  Just wrap the whole set of tests
in a single call to the newly added OVN_FOR_EACH_NORTHD_NO_HV().

This also fixes the fact that some tests were not run for all flavors of
northd.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>