ci: ovn-kubernetes: Pin ovn-k8s version.
Recently, ovn-k8s started using the new OVN Northbound database table
Chassis_Template_Var that was added in v22.12.0.  Unfortunately,
ovn-k8s (libovsdb) became incompatible with OVN 22.09 as it doesn't
support connecting to servers that don't contain this table in their
schema.

It's still important to check that we don't introduce regressions on
stable branches so, instead of completely removing the ovn-k8s CI jobs,
we now pin ovn-k8s to a version that's known to work with OVN 22.09.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>