if-status: track interfaces for additional chassis
This will allow all chassis hosting a port to extract interface MTU from
if-status-mgr. This will be used in a later patch to calculate the
effective path MTU for each port.

In addition, it's the right thing to do to claim and mark an interface
on all chassis as ovn-installed, even if the chassis is "additional".

Fixes: fa8c591fa2a7 ("Support LSP:options:requested-chassis as a list")
Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>