ovn-controller: Assume well-known tables are in the SB schema.
It's safe to assume that tables that existed in the previous LTS branch
first release (currently 22.03.0) can be monitored directly.  Do so and
only "optionally" monitor the ones that have been added since.

This way we avoid the need for the IDL to expose an API to change the
default condition for monitored tables.  It also avoids complex code in
ovn-controller because we'd otherwise have to explicitly re-initialize
conditions to a non-default (false) value after every SB reconnect.

NOTE: we make an exception for the Port_Binding table because, with
conditional monitoring enabled, the condition flap would cause
unnecessary churn in the set of local datapaths and disrupt traffic,
partially reverting the effect of 5a1d82cb28c5 ("controller: Migrate
from ct zone UUID name to component name").

Reported-at: https://mail.openvswitch.org/pipermail/ovs-dev/2023-July/406201.html
Reported-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>