py-requirements: Remove hacking dependency and use recent flake8.
Without this, when using Python 3.12 and flake8 5.0.4, the following
errors are flagged:
  tests/check_acl_log.py:97:25: E231 missing whitespace after ':'
  tests/check_acl_log.py:102:71: E231 missing whitespace after ':'

This was reported and discussed in a couple of contexts:
https://mail.openvswitch.org/pipermail/ovs-dev/2023-November/409325.html
https://mail.openvswitch.org/pipermail/ovs-dev/2023-November/409277.html

And it's fixed in recent flake8/pycodestyle versions:
https://github.com/PyCQA/flake8/issues/1845#issuecomment-1766073353

Unfortunately we have to remove the 'hacking' requirement because that
introduces a dependency on 'flake8<4.0.0 and >=3.6.0'.  On the other
hand the currently enabled hacking checks were only applicable to
Python 2 code.  OVN has stopped supporting Python 2 for a while now,
since 0c042c2d28d8 ("Require Python 3 and remove support for Python
2.").

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
---
NOTE: this patch should be backported to all supported branches.

V3:
- Addressed Ilya's comments:
  - keep hacking checks on flake8 ignore list for the case when hacking
    is installed.

V2:
- Addressed Ilya's comments:
  - removed remaining hacking references
- updated commit log