Documentation: Use theme from Read The Docs.
'ovs-sphinx-theme' is designed to look like an openvswitch.org
website.  It contains OVS logo and navigation bars from the
openvswitch.org.  And that doesn't really make a lot of sense
for OVN.  Also, currently the ovs-sphinx-theme is not actually
installed by the Read The Docs configuration, so the docs.ovn.org
is using default alabaster theme instead.

Switch to sphinx_rtd_theme, it looks close to the main ovn.org
website.  Remove the upper limit on sphinx version, because the
theme may require higher versions and also sphinx 2.0 is very old
and fails to be installed on Read The Docs servers.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>