actions: Remove ct_commit_v1.
It's not used by any logical flow since v20.12.0.  It was kept for
backwards compatibility reasons in 4799f73fd6ed ("Fix OVN update
issue when ovn-controller is updated first from 20.06 to 20.09.") but
now there's no supported release that uses it.

Therefore the action is safe to be removed.  Update the tests to reflect
the fact that the action is not supported anymore.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>