northd: Assume all chassis support the "ct-no-masked-label" feature.
This feature is supported in the last two LTS releases and the correct
upgrade procedure mandates that we don't jump across LTS releases.  It's
safe to remove the check in northd.

Also remove logical field definitions for ct_label.<field>s that cannot
appear anymore in the southbound database.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
---
Changes in v2:
- fixed "Load balancer CT related backwards compatibility" test.