controller, controller-vtep: Add --unixctl option.
All other OVN daemons have the --unixctl option to specify the control
socket file. It was missing from ovn-controller and ovn-controller-vtep,
which this patch fixes.

The ovn-controller-vtep documentation was also missing the already
supported --help and --version options, which are now reflected in the
docs.

Reported-at: https://issues.redhat.com/browse/FDP-714
Signed-off-by: Rosemarie O'Riorden <roriorden@redhat.com>